Rename README.md to T-Tests on Suspension Coils
The 'Total_Summary' for lots 1 through 3 reveals the mean is 1498.78 in all 150 rows. The variance is at 62.29356 for all 150 rows and the standard deviation is at 7.892627 for all 150 rows. The median value of the table varies between 1452 and 1542. The variance of the PSI sample distribution and the standard deviation are within the design specifications for all 3 lots and does not exceed 100 pounds per square inch.