Create ## Study Design: MechaCar vs Competition
Metrics that interest people trying to buy a new car are gas efficiency, horsepower, and safety rating. The tests we performed enable us to compare MechaCar to the the competition.  Our null hypothesis is that MechaCar is not different from the competition.  We can test this by using our t-test after collecting data from different types of competitor vehicles and comparing them to the data cobtained for MechaCar.